+++ src/RcppExports.cpp
@@ -74,6 +74,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// _dump_open_datasets
+int _dump_open_datasets(std::string outfile);
+RcppExport SEXP _gdalraster__dump_open_datasets(SEXP outfileSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< std::string >::type outfile(outfileSEXP);
+    rcpp_result_gen = Rcpp::wrap(_dump_open_datasets(outfile));
+    return rcpp_result_gen;
+END_RCPP
+}
 // push_error_handler
 void push_error_handler(std::string handler);
 RcppExport SEXP _gdalraster_push_error_handler(SEXP handlerSEXP) {
@@ -142,6 +153,15 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     rcpp_result_gen = Rcpp::wrap(http_enabled());
     return rcpp_result_gen;
+END_RCPP
+}
+// _cpl_http_cleanup
+void _cpl_http_cleanup();
+RcppExport SEXP _gdalraster__cpl_http_cleanup() {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    _cpl_http_cleanup();
+    return R_NilValue;
 END_RCPP
 }
 // create
@@ -1444,6 +1464,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_gdalraster_get_config_option", (DL_FUNC) &_gdalraster_get_config_option, 1},
     {"_gdalraster_set_config_option", (DL_FUNC) &_gdalraster_set_config_option, 2},
     {"_gdalraster_get_cache_used", (DL_FUNC) &_gdalraster_get_cache_used, 0},
+    {"_gdalraster__dump_open_datasets", (DL_FUNC) &_gdalraster__dump_open_datasets, 1},
     {"_gdalraster_push_error_handler", (DL_FUNC) &_gdalraster_push_error_handler, 1},
     {"_gdalraster_pop_error_handler", (DL_FUNC) &_gdalraster_pop_error_handler, 0},
     {"_gdalraster__check_gdal_filename", (DL_FUNC) &_gdalraster__check_gdal_filename, 1},
@@ -1451,6 +1472,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_gdalraster_get_usable_physical_ram", (DL_FUNC) &_gdalraster_get_usable_physical_ram, 0},
     {"_gdalraster_has_spatialite", (DL_FUNC) &_gdalraster_has_spatialite, 0},
     {"_gdalraster_http_enabled", (DL_FUNC) &_gdalraster_http_enabled, 0},
+    {"_gdalraster__cpl_http_cleanup", (DL_FUNC) &_gdalraster__cpl_http_cleanup, 0},
     {"_gdalraster_create", (DL_FUNC) &_gdalraster_create, 7},
     {"_gdalraster_createCopy", (DL_FUNC) &_gdalraster_createCopy, 6},
     {"_gdalraster__apply_geotransform", (DL_FUNC) &_gdalraster__apply_geotransform, 3},